v1.3.1: Flock-You WiFi sniffer fixes — addr1 matching, wildcard probe, RSSI floor, FCS retry
- addr1 (receiver) OUI matching: sleeping Flock cameras appear as
  destinations of probe responses; checking only addr2 missed them.
- Wildcard probe detection: empty SSID (tag 0, len 0) from known
  OUI = METHOD_WILDCARD_PROBE at 65 pts, bypasses OUI-only cap.
- RSSI floor: drop frames below -95 dBm in callback to keep ring
  buffer clean.
- FCS trailer retry: some ESP32 drivers include 4-byte FCS in
  sig_len; retry SSID IE parsing with body_len-4 if first pass
  found nothing.
- Frame subtype encoded into frame_type byte (hi nibble=type,
  lo nibble=subtype) without changing struct size.
- SSID substring matching: indexOfIgnoreCase replaces broken
  position-0 char-by-char comparison for camera/carrier SSIDs.
- methods field expanded u16→u32 (all 16 bits consumed).

+++ src/boards/heltec_v3_ui.zig
@@ -111,7 +111,7 @@ fn formatMac(mac: [6]u8, buf: []u8) []const u8 {
 }
 
 /// Raven firmware version string from methods flags.
-fn ravenFwStr(methods: u16) []const u8 {
+fn ravenFwStr(methods: u32) []const u8 {
     if (methods & scanner.RAVEN_FW_1_3 != 0) return "v1.3";
     if (methods & scanner.RAVEN_FW_1_2 != 0) return "v1.2";
     if (methods & scanner.RAVEN_FW_1_1 != 0) return "v1.1";
